MiniConsoles have a search bar suddenly #8275
